CheckGroup.js moved folder and imports corrected
