+++ src/components/atoms/Checkbox.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { useField } from "formik";
 import styled from "styled-components";
-import CheckGroup from '../molecules/CheckGroup';
+import CheckGroup from '../atoms/CheckGroup';
 import Label from "./Label";
 
 const StyledCheckboxField = styled.input`
@@ -10,12 +10,12 @@ const StyledCheckboxField = styled.input`
   margin: 0 20px 10px 0;
 `;
 
-function Checkbox({label, ...props}) {
+function Checkbox({ label, ...props }) {
   const [field] = useField(props);
 
   return (
     <CheckGroup>
-      <StyledCheckboxField id={props.value} type="checkbox" {...field} {...props}/>
+      <StyledCheckboxField id={props.value} type="checkbox" {...field} {...props} />
       <Label htmlFor={props.value}>{label}</Label>
     </CheckGroup>
   );
